+++ elenasrc2/engine/win64/win64routines.cpp
@@ -1,7 +1,7 @@
-//---------------------------------------------------------------------------
+//----------------------------------------------------------------------r-----
 //		E L E N A   P r o j e c t:  Win64 ELENA System Routines
 //
-//                                              (C)2020, by Alexei Rakov
+//                                              (C)2020-2021, by Alexei Rakov
 //---------------------------------------------------------------------------
 
 #include "elena.h"
@@ -11,20 +11,66 @@
 
 using namespace _ELENA_;
 
-void SystemRoutineProvider::InitCriticalStruct(CriticalStruct* header, uintptr_t criticalHandler)
+static uintptr_t CriticalHandler = 0;
+
+#define CALL_FIRST 1  
+
+void SystemRoutineProvider::RaiseError(int code)
 {
-   //uintptr_t previousHeader = 0;
+   ::RaiseException(code, 0, 0, 0);
+}
 
-   //// ; set SEH handler / frame / stack pointers
-   //__asm {
-   //   mov  rax, header
-   //   mov  rcx, fs: [0]
-   //   mov  previousHeader, rcx
-   //   mov  fs : [0] , rax
-   //}
+intptr_t SystemRoutineProvider::AlignHeapSize(intptr_t size)
+{
+   return align(size, 128);
+}
 
-   //header->previousStruct = previousHeader;
-   //header->handler = criticalHandler;
+LONG WINAPI ELENAVectoredHandler(struct _EXCEPTION_POINTERS* ExceptionInfo)
+{
+   int r = 0;
+   switch (ExceptionInfo->ExceptionRecord->ExceptionCode)
+   {
+      case EXCEPTION_BREAKPOINT:
+      case EXCEPTION_SINGLE_STEP:
+      case EXCEPTION_NONCONTINUABLE_EXCEPTION:
+      case CONTROL_C_EXIT:
+         return EXCEPTION_CONTINUE_SEARCH;
+      case EXCEPTION_ACCESS_VIOLATION:
+         ExceptionInfo->ContextRecord->Rdx = ExceptionInfo->ContextRecord->Rip;
+         ExceptionInfo->ContextRecord->Rax = ELENA_ERR_ACCESS_VIOLATION;
+         ExceptionInfo->ContextRecord->Rip = CriticalHandler;
+
+         return EXCEPTION_CONTINUE_EXECUTION;
+      case ELENA_ERR_OUT_OF_MEMORY:
+         ExceptionInfo->ContextRecord->Rdx = ExceptionInfo->ContextRecord->Rip;
+         ExceptionInfo->ContextRecord->Rax = ELENA_ERR_OUT_OF_MEMORY;
+         ExceptionInfo->ContextRecord->Rip = CriticalHandler;
+         ExceptionInfo->ContextRecord->Rbx = 0;
+
+         return EXCEPTION_CONTINUE_EXECUTION;
+      case EXCEPTION_INT_DIVIDE_BY_ZERO:
+      case EXCEPTION_FLT_DIVIDE_BY_ZERO:
+         ExceptionInfo->ContextRecord->Rdx = ExceptionInfo->ContextRecord->Rip;
+         ExceptionInfo->ContextRecord->Rax = ELENA_ERR_DIVIDE_BY_ZERO;
+         ExceptionInfo->ContextRecord->Rip = CriticalHandler;
+
+         return EXCEPTION_CONTINUE_EXECUTION;
+      default:
+         ExceptionInfo->ContextRecord->Rdx = ExceptionInfo->ContextRecord->Rip;
+         ExceptionInfo->ContextRecord->Rax = ELENA_ERR_CRITICAL;
+         ExceptionInfo->ContextRecord->Rip = CriticalHandler;
+
+         return EXCEPTION_CONTINUE_EXECUTION;
+   }
+
+   return EXCEPTION_CONTINUE_SEARCH;
+}
+
+void SystemRoutineProvider::InitCriticalStruct(uintptr_t criticalHandler)
+{
+   CriticalHandler = criticalHandler;
+
+   AddVectoredExceptionHandler(CALL_FIRST, ELENAVectoredHandler);
 }
 
 void SystemRoutineProvider::InitTLSEntry(pos_t threadIndex, pos_t tlsIndex, ProgramHeader* frameHeader, pos_t* threadTable)